dissolve cfunction.py
