+++ firedrake/__init__.py
@@ -31,7 +31,6 @@ from pyop2 import op2                                           # noqa
 from firedrake.assemble import *
 from firedrake.bcs import *
 from firedrake.constant import *
-from firedrake.cfunction import *
 from firedrake.expression import *
 from firedrake.function import *
 from firedrake.functionspace import *
